HBASE-29383 Some tests in TestHFileOutputFormat2 fails in ShutdownHook.suppressHdfsShutdownHook
Split TestHFileOutputFormat2 to make it run faster and prevent the tests
affect each other
